+++ src/app/plugins.js
@@ -6,18 +6,28 @@ import {
   DEFAULT_COORDINATION_VALUES,
 } from './state/coordination';
 
-export const PLUGIN_VIEW_TYPES_KEY = '__VITESSCE_PLUGIN_VIEW_TYPES__';
-export const PLUGIN_COORDINATION_TYPES_KEY = '__VITESSCE_PLUGIN_COORDINATION_TYPES__';
-export const PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY = '__VITESSCE_PLUGIN_COORDINATION_TYPES_PER_VIEW__';
-export const PLUGIN_FILE_TYPES_KEY = '__VITESSCE_PLUGIN_FILE_TYPES__';
+const PLUGINS_KEY = '__VITESSCE_PLUGINS__';
+const PLUGIN_VIEW_TYPES_KEY = 'viewTypes';
+const PLUGIN_COORDINATION_TYPES_KEY = 'coordinationTypes';
+const PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY = 'coordinationTypesPerView';
+const PLUGIN_FILE_TYPES_KEY = 'fileTypes';
+const PLUGIN_FILE_TYPE_DATA_TYPE_MAPPING_KEY = 'fileTypeDataTypeMapping';
+const PLUGIN_CONVENIENCE_FILE_TYPES_KEY = 'convenienceFileTypes';
+
+window[PLUGINS_KEY] = {};
+const PLUGINS = window[PLUGINS_KEY];
 
 // Reference: https://github.com/higlass/higlass-register/blob/master/src/index.js
-window[PLUGIN_VIEW_TYPES_KEY] = window[PLUGIN_VIEW_TYPES_KEY] || {};
-window[PLUGIN_COORDINATION_TYPES_KEY] = window[PLUGIN_COORDINATION_TYPES_KEY] || {};
-window[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY] = (
-  window[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY] || {}
+PLUGINS[PLUGIN_VIEW_TYPES_KEY] = PLUGINS[PLUGIN_VIEW_TYPES_KEY] || {};
+PLUGINS[PLUGIN_COORDINATION_TYPES_KEY] = PLUGINS[PLUGIN_COORDINATION_TYPES_KEY] || {};
+PLUGINS[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY] = (
+  PLUGINS[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY] || {}
 );
-window[PLUGIN_FILE_TYPES_KEY] = window[PLUGIN_FILE_TYPES_KEY] || {};
+PLUGINS[PLUGIN_FILE_TYPES_KEY] = PLUGINS[PLUGIN_FILE_TYPES_KEY] || {};
+PLUGINS[PLUGIN_FILE_TYPE_DATA_TYPE_MAPPING_KEY] = (
+  PLUGINS[PLUGIN_FILE_TYPE_DATA_TYPE_MAPPING_KEY] || {}
+);
+PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY] = PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY] || {};
 
 /**
  * Register a new coordination type.
@@ -25,7 +35,7 @@ window[PLUGIN_FILE_TYPES_KEY] = window[PLUGIN_FILE_TYPES_KEY] || {};
  * @param {*} defaultValue A default value for the coordination type.
  */
 export function registerPluginCoordinationType(typeName, defaultValue) {
-  window[PLUGIN_COORDINATION_TYPES_KEY][typeName] = defaultValue;
+  PLUGINS[PLUGIN_COORDINATION_TYPES_KEY][typeName] = defaultValue;
 }
 
 /**
@@ -35,9 +45,9 @@ export function registerPluginCoordinationType(typeName, defaultValue) {
  * @param {string[]} coordinationTypes A list of coordination types that this view supports.
  */
 export function registerPluginViewType(viewType, viewSubscriberReactComponent, coordinationTypes) {
-  window[PLUGIN_VIEW_TYPES_KEY][viewType] = viewSubscriberReactComponent;
+  PLUGINS[PLUGIN_VIEW_TYPES_KEY][viewType] = viewSubscriberReactComponent;
   // Register the supported coordination types.
-  const pluginTypesPerView = window[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY];
+  const pluginTypesPerView = PLUGINS[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY];
   if (Array.isArray(pluginTypesPerView[viewType])) {
     coordinationTypes.forEach((coordinationType) => {
       if (!pluginTypesPerView[viewType].includes(coordinationType)) {
@@ -66,41 +76,61 @@ export function registerPluginFileType(
   };
 }
 
+/**
+ * Register a new file type.
+ * @param {string} fileTypeName Name for the new file type.
+ * @param {function} expansionFunction The file type expansion function.
+ * Should take in a single file definition and return an array of
+ * file definitions with valid fileType values.
+ */
+export function registerPluginConvenienceFileType(
+  // eslint-disable-next-line no-unused-vars
+  fileTypeName, expansionFunction,
+) {
+  PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY][fileTypeName] = expansionFunction;
+}
+
 
 // Plugin getter functions.
 
 export function getPluginViewTypes() {
-  return Object.keys(window[PLUGIN_VIEW_TYPES_KEY]);
+  return Object.keys(PLUGINS[PLUGIN_VIEW_TYPES_KEY]);
 }
 
 export function getPluginViewType(viewType) {
-  return window[PLUGIN_VIEW_TYPES_KEY][viewType];
+  return PLUGINS[PLUGIN_VIEW_TYPES_KEY][viewType];
 }
 
 export function getPluginCoordinationTypes() {
-  return Object.keys(window[PLUGIN_COORDINATION_TYPES_KEY]);
+  return Object.keys(PLUGINS[PLUGIN_COORDINATION_TYPES_KEY]);
 }
 
 export function getPluginCoordinationTypeDefaults() {
-  return window[PLUGIN_COORDINATION_TYPES_KEY];
+  return PLUGINS[PLUGIN_COORDINATION_TYPES_KEY];
 }
 
 export function getPluginCoordinationTypesForViewType(viewType) {
-  if (Array.isArray(window[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY][viewType])) {
-    return window[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY][viewType];
+  if (Array.isArray(PLUGINS[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY][viewType])) {
+    return PLUGINS[PLUGIN_COORDINATION_TYPES_PER_VIEW_KEY][viewType];
   }
   return [];
 }
 
 export function getPluginFileTypes() {
-  return Object.keys(window[PLUGIN_FILE_TYPES_KEY]);
+  return Object.keys(PLUGINS[PLUGIN_FILE_TYPES_KEY]);
+}
+export function getPluginConvenienceFileTypes() {
+  return Object.keys(PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY]);
 }
 
 export function getDataTypeForPluginFileType(fileType) {
-  return window[PLUGIN_FILE_TYPES_KEY][fileType].dataType;
+  return PLUGINS[PLUGIN_FILE_TYPE_DATA_TYPE_MAPPING_KEY][fileType];
 }
 export function getLoaderClassesForPluginFileType(fileType) {
-  return window[PLUGIN_FILE_TYPES_KEY][fileType].loaderClasses;
+  return PLUGINS[PLUGIN_FILE_TYPES_KEY][fileType];
+}
+export function getExpansionFunctionForPluginConvenienceFileType(fileType) {
+  return PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY][fileType];
 }
 
 // Getters that depend on plugins.
@@ -146,5 +176,12 @@ export function getFileTypeDataTypeMapping() {
       fileType,
       getDataTypeForPluginFileType(fileType),
     ]))),
+  };
+}
+
+export function getConvenienceFileTypes() {
+  return {
+    // TODO: import built-in convenience file types and include them here.
+    ...PLUGINS[PLUGIN_CONVENIENCE_FILE_TYPES_KEY],
   };
 }